chore: move handler to internal
